# (maint) Speed up rspec tests