@puppetcla 2020-01-28
CLA signed by all contributors.